Updated graphics module, 2D renderer and Sprite Editor
Graphics:  Issue with setting multiple render targets and then setting a single target - The validation code kicked in too early and triggered an exception because old targets were still bound before the target list was updated.

2D Renderer:
   GorgonSprite/HLSL code - Added an attribute to the sprite vertex declaration to include the angle of rotation for the sprite.  This allows easier (and more efficient) normal mapping.  Normal maps need to be transformed by the world transform in order to be correct when calculating lighting.

Gorgon Editor:  Updated the sprite editor to allow clipping a sprite from an image just by clicking on it.  Uses a flood fill algorithm to determine the boundaries of the sprite and defines its rectangle from those boundaries.

+++ Tools/Gorgon.Editor.PlugIns/Gorgon.SpriteEditor/AutoClipper.cs
@@ -137,7 +137,7 @@ namespace GorgonLibrary.Editor.SpriteEditorPlugIn
 			{
 				case BufferFormat.B8G8R8A8_UIntNormal:
 				case BufferFormat.B8G8R8A8_UIntNormal_sRGB:
-					encodedValue = (uint)((UseAlpha ? (uint)(MaskColor.A << 24) : 0xff000000) | (MaskColor.R << 16) | (MaskColor.G << 8) | (MaskColor.B));
+					encodedValue = ((UseAlpha ? ((uint)MaskColor.A << 24) : 0xff000000) | ((uint)MaskColor.R << 16) | ((uint)MaskColor.G << 8) | (MaskColor.B));
 					pixelValue = *((uint*)pixelPtr);
 
 					if (!UseAlpha)
@@ -150,7 +150,7 @@ namespace GorgonLibrary.Editor.SpriteEditorPlugIn
 				case BufferFormat.R8G8B8A8_UIntNormal_sRGB:
 				case BufferFormat.R8G8B8A8_IntNormal:
 				case BufferFormat.R8G8B8A8_Int:
-					encodedValue = (uint)((UseAlpha ? (uint)(MaskColor.A << 24) : 0xff000000) | (MaskColor.B << 16) | (MaskColor.G << 8) | (MaskColor.R));
+					encodedValue = ((UseAlpha ? ((uint)MaskColor.A << 24) : 0xff000000) | ((uint)MaskColor.B << 16) | ((uint)MaskColor.G << 8) | (MaskColor.R));
 					pixelValue = *((uint*)pixelPtr);
 
 					if (!UseAlpha)
